Rename layout to Layout.js, remove image.js
