+++ src/pages/index.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
-import Image from '../components/image'
+import Layout from '../components/Layout'
 
 const IndexPage = () => (
   <Layout>
