+++ src/pages/page-2.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
+import Layout from '../components/Layout'
 
 const SecondPage = () => (
   <Layout>
